feat: add basic lexer
okay so this is properly lexing the add sample, but i need to do some more qol
stuff in the lexer. for example, i want to be able to specify some tokens should
be skipped, so that they won't show up in the result token type enum. also need
to fix up the error messages around errors. probably also going to overhaul the
api around the lexer

Signed-off-by: Matt Provost <mattprovost6@gmail.com>